+++ BACnetServerExample/BACnetServerExample.cpp
@@ -63,7 +63,7 @@ time_t g_warmStartTimer; // Timer used for delaying the warm start.
 
 // Constants
 // =======================================
-const std::string APPLICATION_VERSION = "0.0.25";  // See CHANGELOG.md for a full list of changes.
+const std::string APPLICATION_VERSION = "0.0.26";  // See CHANGELOG.md for a full list of changes.
 const uint32_t MAX_RENDER_BUFFER_LENGTH = 1024 * 20;
 
 
@@ -464,9 +464,10 @@ bool SetupDevice() {
 	fpSetPropertySubscribable(g_exampleDatabase.device.instance, CASBACnetStackExampleConstants::OBJECT_TYPE_ANALOG_INPUT, g_exampleDatabase.analogInput.instance, CASBACnetStackExampleConstants::PROPERTY_IDENTIFIER_PRESENT_VALUE, true);
 	fpSetPropertyWritable(g_exampleDatabase.device.instance, CASBACnetStackExampleConstants::OBJECT_TYPE_ANALOG_INPUT, g_exampleDatabase.analogInput.instance, CASBACnetStackExampleConstants::PROPERTY_IDENTIFIER_COV_INCURMENT, true);
 
-	// Enable the description, and Reliability property 
+	// Enable the description, units and Reliability property 
 	fpSetPropertyByObjectTypeEnabled(g_exampleDatabase.device.instance, CASBACnetStackExampleConstants::OBJECT_TYPE_ANALOG_INPUT, CASBACnetStackExampleConstants::PROPERTY_IDENTIFIER_DESCRIPTION, true);
 	fpSetPropertyByObjectTypeEnabled(g_exampleDatabase.device.instance, CASBACnetStackExampleConstants::OBJECT_TYPE_ANALOG_INPUT, CASBACnetStackExampleConstants::PROPERTY_IDENTIFIER_RELIABILITY, true);
+	fpSetPropertyByObjectTypeEnabled(g_exampleDatabase.device.instance, CASBACnetStackExampleConstants::OBJECT_TYPE_ANALOG_INPUT, CASBACnetStackExampleConstants::PROPERTY_IDENTIFIER_UNITS, true);
 
 	// Enable a specific property to be subscribable for COVProperty 
 	fpSetPropertySubscribable(g_exampleDatabase.device.instance, CASBACnetStackExampleConstants::OBJECT_TYPE_ANALOG_INPUT, g_exampleDatabase.analogInput.instance, CASBACnetStackExampleConstants::PROPERTY_IDENTIFIER_RELIABILITY, true);
@@ -1119,10 +1120,19 @@ bool CallbackGetPropertyCharString(const uint32_t deviceInstance, const uint16_t
 	}
 	// Example of Device Desription
 	else if (propertyIdentifier == CASBACnetStackExampleConstants::PROPERTY_IDENTIFIER_DESCRIPTION) {
-		if (g_exampleDatabase.device.description.size() <= maxElementCount) {
-			memcpy(value, g_exampleDatabase.device.description.c_str(), g_exampleDatabase.device.description.size());
-			*valueElementCount = (uint32_t)g_exampleDatabase.device.description.size();
-			return true;
+		if (objectType == CASBACnetStackExampleConstants::OBJECT_TYPE_DEVICE && objectInstance == g_exampleDatabase.device.instance) {
+			if (g_exampleDatabase.device.description.size() <= maxElementCount) {
+				memcpy(value, g_exampleDatabase.device.description.c_str(), g_exampleDatabase.device.description.size());
+				*valueElementCount = (uint32_t)g_exampleDatabase.device.description.size();
+				return true;
+			}
+		}
+		else if (objectType == CASBACnetStackExampleConstants::OBJECT_TYPE_ANALOG_INPUT && objectInstance == g_exampleDatabase.analogInput.instance) {
+			if (g_exampleDatabase.analogInput.description.size() <= maxElementCount) {
+				memcpy(value, g_exampleDatabase.analogInput.description.c_str(), g_exampleDatabase.analogInput.description.size());
+				*valueElementCount = (uint32_t)g_exampleDatabase.analogInput.description.size();
+				return true;
+			}
 		}
 		return false;
 	}
@@ -1304,6 +1314,12 @@ bool CallbackGetPropertyEnum(uint32_t deviceInstance, uint16_t objectType, uint3
 	else if (propertyIdentifier == CASBACnetStackExampleConstants::PROPERTY_IDENTIFIER_FD_BBMD_ADDRESS) {
 		if (objectType == CASBACnetStackExampleConstants::OBJECT_TYPE_NETWORK_PORT && objectInstance == g_exampleDatabase.networkPort.instance) {
 			*value = g_exampleDatabase.networkPort.FdBbmdAddressHostType;
+			return true;
+		}
+	}
+	else if (propertyIdentifier == CASBACnetStackExampleConstants::PROPERTY_IDENTIFIER_UNITS) {
+		if (objectType == CASBACnetStackExampleConstants::OBJECT_TYPE_ANALOG_INPUT && objectInstance == g_exampleDatabase.analogInput.instance) {
+			*value = g_exampleDatabase.analogInput.units;
 			return true;
 		}
 	}